Shenanigans (#13)
* swap out Gin fork

* swapped Gin, moved JUCE

* update JUCE

* shenanigans

* m

* knobs

* knob colors

* inactive knobs

* ladder

* create params

* ladder params

* shenan agains

+++ Source/APSamplerVoice.h
@@ -75,7 +75,7 @@ private:
 	double pitchStride;
 	juce::ADSR adsr;
     
-	double sourceSamplePosition = 0;
+	float sourceSamplePosition = 0;
 	float lgain = 0, rgain = 0;
 
 	bool hasLooped{ false };
